b64 PSBT from SD card: refactor and more tests
